+++ interview/api/users-api/src/main/java/com/jonioliveira/users/resource/model/response/UserResponse.java
@@ -6,23 +6,23 @@ import org.eclipse.microprofile.openapi.annotations.media.Schema;
 @Schema(name="User Response Model", description="Model of response body when login users")
 public class UserResponse {
     @Schema(description = "Id of user", example = "1", type = SchemaType.NUMBER)
-    private long id;
+    private int id;
     @Schema(description = "Name of user", example = "jhon", type = SchemaType.STRING)
     private String name;
     @Schema(description = "Type of user", example = "1", type = SchemaType.NUMBER)
-    private long type;
+    private int type;
 
-    public UserResponse(long id, String name, long type) {
+    public UserResponse(int id, String name, int type) {
         this.id = id;
         this.name = name;
         this.type = type;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
@@ -34,11 +34,11 @@ public class UserResponse {
         this.name = name;
     }
 
-    public long getType() {
+    public int getType() {
         return type;
     }
 
-    public void setType(long type) {
+    public void setType(int type) {
         this.type = type;
     }
 }
